+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -93,6 +93,7 @@ import com.vmware.xenon.common.jwt.Verifier;
 import com.vmware.xenon.services.common.AuthCredentialsService;
 import com.vmware.xenon.services.common.AuthorizationContextService;
 import com.vmware.xenon.services.common.AuthorizationTokenCacheService;
+import com.vmware.xenon.services.common.CheckpointFactoryService;
 import com.vmware.xenon.services.common.ConsistentHashingNodeSelectorService;
 import com.vmware.xenon.services.common.DirectoryContentService;
 import com.vmware.xenon.services.common.GraphQueryTaskService;
@@ -1640,6 +1641,12 @@ public class ServiceHost implements ServiceRequestSender {
 
             }
         }
+
+        // check point depends on index service
+        // synchronization task service may lookup check point
+        CheckpointFactoryService service = new CheckpointFactoryService();
+        service.setUseBodyForSelfLink(true);
+        startCoreServicesSynchronously(service);
 
         List<Service> coreServices = new ArrayList<>();
         coreServices.add(this.managementService);