+++ xenon-common/src/main/java/com/vmware/xenon/common/SynchronizationTaskService.java
@@ -17,12 +17,18 @@ import java.net.URI;
 import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
 
+import com.vmware.xenon.common.NodeSelectorService.SelectOwnerResponse;
 import com.vmware.xenon.common.ServiceDocumentDescription.PropertyUsageOption;
+import com.vmware.xenon.common.config.XenonConfiguration;
+import com.vmware.xenon.services.common.CheckpointService;
+import com.vmware.xenon.services.common.NodeGroupBroadcastResponse;
 import com.vmware.xenon.services.common.QueryTask;
 import com.vmware.xenon.services.common.QueryTask.QuerySpecification.QueryOption;
 import com.vmware.xenon.services.common.ServiceUriPaths;
@@ -63,7 +69,7 @@ public class SynchronizationTaskService
     }
 
     public enum SubStage {
-        QUERY, SYNCHRONIZE, RESTART, CHECK_NG_AVAILABILITY
+        GET_CHECKPOINTS, QUERY, SYNCHRONIZE, RESTART, CHECK_NG_AVAILABILITY
     }
 
     public static class State extends TaskService.TaskServiceState {
@@ -109,6 +115,17 @@ public class SynchronizationTaskService
         public Long membershipUpdateTimeMicros;
 
         /**
+         * start time of the synchronization state machine
+         */
+        public Long startTimeMicros;
+
+        /**
+         * The last known time that peer synced given factory and node group
+         */
+        @UsageOption(option = PropertyUsageOption.AUTO_MERGE_IF_NOT_NULL)
+        public Long checkpoint;
+
+        /**
          * The current SubStage of the synchronization task.
          */
         @UsageOption(option = PropertyUsageOption.AUTO_MERGE_IF_NOT_NULL)
@@ -129,8 +146,22 @@ public class SynchronizationTaskService
 
     private Supplier<Service> childServiceInstantiator;
 
+    private FactoryService parent;
+
     private final boolean isDetailedLoggingEnabled = Boolean
             .getBoolean(PROPERTY_NAME_SYNCHRONIZATION_LOGGING);
+
+    private final boolean isCheckpointEnabled = XenonConfiguration.bool(
+            SynchronizationTaskService.class,
+            "isCheckpointEnabled",
+            true
+    );
+
+    private final long schedulePeriodSeconds = XenonConfiguration.number(
+            SynchronizationTaskService.class,
+            "schedulePeriodSeconds",
+            TimeUnit.MINUTES.toSeconds(30)
+    );
 
     public SynchronizationTaskService() {
         super(State.class);
@@ -178,6 +209,7 @@ public class SynchronizationTaskService
         initialState.childOptions = childTemplate.getOptions();
         initialState.childDocumentIndexLink = childTemplate.getDocumentIndexPath();
         initialState.documentExpirationTimeMicros = Long.MAX_VALUE;
+        initialState.checkpoint = 0L;
     }
 
     @Override
@@ -287,7 +319,12 @@ public class SynchronizationTaskService
         task.queryResultLimit = body.queryResultLimit;
         if (startStateMachine) {
             task.taskInfo.stage = TaskState.TaskStage.STARTED;
-            task.subStage = SubStage.QUERY;
+            if (this.parent != null && this.parent.hasChildOption(ServiceOption.PERSISTENCE) && this.isCheckpointEnabled) {
+                task.subStage = SubStage.GET_CHECKPOINTS;
+            } else {
+                task.subStage = SubStage.QUERY;
+                task.checkpoint = 0L;
+            }
         }
 
         if (this.isDetailedLoggingEnabled) {
@@ -304,6 +341,7 @@ public class SynchronizationTaskService
             // redundant since the FactoryService may already have
             // changed the status to un-available, but just for
             // correctness we do it here again.
+            task.startTimeMicros = Utils.getNowMicrosUtc();
             task.synchCompletionCount = 0;
             setStat(STAT_NAME_CHILD_SYNCH_RETRY_COUNT, 0);
             setStat(STAT_NAME_CHILD_SYNCH_FAILURE_COUNT, 0);
@@ -399,7 +437,12 @@ public class SynchronizationTaskService
             // the task's stage to RESTART. In this case, we reset the task
             // back to QUERY stage.
             task.taskInfo.stage = TaskState.TaskStage.STARTED;
-            task.subStage = SubStage.QUERY;
+            if (this.parent != null && this.parent.hasChildOption(ServiceOption.PERSISTENCE) && this.isCheckpointEnabled) {
+                task.subStage = SubStage.GET_CHECKPOINTS;
+            } else {
+                task.subStage = SubStage.QUERY;
+                task.checkpoint = 0L;
+            }
             task.synchCompletionCount = 0;
             setStat(STAT_NAME_CHILD_SYNCH_RETRY_COUNT, 0);
             setStat(STAT_NAME_CHILD_SYNCH_FAILURE_COUNT, 0);
@@ -441,6 +484,9 @@ public class SynchronizationTaskService
 
     public void handleSubStage(State task) {
         switch (task.subStage) {
+        case GET_CHECKPOINTS:
+            handleCheckpointStage(task);
+            break;
         case QUERY:
             handleQueryStage(task);
             break;
@@ -454,6 +500,58 @@ public class SynchronizationTaskService
             logWarning("Unexpected sub stage: %s", task.subStage);
             break;
         }
+    }
+
+    private void handleCheckpointStage(State task) {
+        String checkPointServiceLink = UriUtils.buildUriPath(
+                CheckpointService.FACTORY_LINK, UriUtils.convertPathCharsFromLink(this.parent.getSelfLink()));;
+        Operation get = Operation.createGet(UriUtils.buildUri(this.getHost(), checkPointServiceLink))
+                .setReferer(this.getUri())
+                .setCompletion((o, e) -> {
+                    if (e != null) {
+                        logInfo("broadcast get checkpoints failed %s, starting synchronization from timestamp 0", e.toString());
+                        task.checkpoint = 0L;
+                        sendSelfPatch(task, TaskState.TaskStage.STARTED,
+                                subStageSetter(SubStage.QUERY));
+                        return;
+                    }
+                    NodeGroupBroadcastResponse rsp = o.getBody(NodeGroupBroadcastResponse.class);
+                    if (!rsp.failures.isEmpty()) {
+                        for (Map.Entry<URI, ServiceErrorResponse> failure : rsp.failures.entrySet()) {
+                            // 404 may due to checkpoint is not created yet
+                            if (failure.getValue().statusCode != Operation.STATUS_CODE_NOT_FOUND) {
+                                logInfo("get checkpoint failed with status %d from %s", failure.getValue().errorCode, failure.getKey());
+                            }
+                        }
+                        logInfo("starting synchronization from timestamp 0");
+                        task.checkpoint = 0L;
+                        sendSelfPatch(task, TaskState.TaskStage.STARTED,
+                                subStageSetter(SubStage.QUERY));
+                        return;
+                    }
+                    List<Long> checkPoints =
+                            rsp.jsonResponses.values().stream().map(s -> {
+                                CheckpointService.CheckpointState checkpointState =
+                                        Utils.fromJson(s, CheckpointService.CheckpointState.class);
+                                return checkpointState.timestamp;
+                            }).collect(Collectors.toList());
+                    task.checkpoint = findMinimumCheckpoint(checkPoints);
+                    if (task.checkpoint > 0) {
+                        logInfo("synch %s from check point %d",
+                                task.factorySelfLink, task.checkpoint);
+                    }
+                    sendSelfPatch(task, TaskState.TaskStage.STARTED,
+                            subStageSetter(SubStage.QUERY));
+                });
+        this.getHost().broadcastRequest(task.nodeSelectorLink, checkPointServiceLink, false, get);
+    }
+
+    private long findMinimumCheckpoint(List<Long> checkpoints) {
+        long minimumCheckpoint = Long.MAX_VALUE;
+        for (Long checkpoint : checkpoints) {
+            minimumCheckpoint = Long.min(minimumCheckpoint, checkpoint);
+        }
+        return minimumCheckpoint;
     }
 
     private void handleQueryStage(State task) {
@@ -518,6 +616,21 @@ public class SynchronizationTaskService
                 .setTermMatchValue(task.factoryStateKind);
         queryTask.querySpec.query.addBooleanClause(kindClause);
 
+        if (this.parent != null && this.parent.hasChildOption(ServiceOption.PERSISTENCE) && this.isCheckpointEnabled) {
+            if (this.isDetailedLoggingEnabled) {
+                logInfo("query %s from checkpoint %d", task.factorySelfLink, task.checkpoint);
+            }
+
+            QueryTask.NumericRange<Long> timeRange =
+                    QueryTask.NumericRange.createLongRange(task.checkpoint, Long.MAX_VALUE,
+                            false, true);
+
+            QueryTask.Query timeClause = new QueryTask.Query()
+                    .setTermPropertyName(ServiceDocument.FIELD_NAME_UPDATE_TIME_MICROS)
+                    .setNumericRange(timeRange);
+            queryTask.querySpec.query.addBooleanClause(timeClause);
+        }
+
         // set timeout based on peer synchronization upper limit
         long timeoutMicros = TimeUnit.SECONDS.toMicros(
                 getHost().getPeerSynchronizationTimeLimitSeconds());
@@ -547,9 +660,20 @@ public class SynchronizationTaskService
             return;
         }
 
-        if (verifyOwnership && verifySynchronizationOwnership(task)) {
-            // Verifying ownership will recursively call into
-            // handleSynchronizationStage with verifyOwnership set to false.
+        if (verifyOwnership && task.childOptions.contains(ServiceOption.REPLICATION)) {
+            Consumer<SelectOwnerResponse> ownerHandler = (selectOwnerResponse) -> {
+                handleSynchronizeStage(task, false);
+            };
+            Consumer<SelectOwnerResponse> nonOwnerHandler = (selectOwnerResponse) -> {
+                logWarning("Current node %s is no longer owner for the factory %s. Cancelling synchronization",
+                        this.getHost().getId(), task.factorySelfLink);
+                sendSelfCancellationPatch(task, "Local node is no longer owner for this factory.");
+            };
+            Consumer<Throwable> failureHandler = (e) -> {
+                sendSelfFailurePatch(task, e.getMessage());
+            };
+
+            verifySynchronizationOwnership(task, ownerHandler, nonOwnerHandler, failureHandler);
             return;
         }
 
@@ -738,40 +862,34 @@ public class SynchronizationTaskService
         return delay;
     }
 
-    private boolean verifySynchronizationOwnership(State task) {
-        // If this is not a REPLICATED factory, we don't
-        // bother verifying ownership.
-        if (!task.childOptions.contains(ServiceOption.REPLICATION)) {
-            return false;
-        }
-
+    private void verifySynchronizationOwnership(State task, Consumer<SelectOwnerResponse> ownerHandler,
+            Consumer<SelectOwnerResponse> nonOwnerHandler, Consumer<Throwable> failureHandler) {
         Operation selectOp = Operation
                 .createPost(null)
                 .setExpiration(task.documentExpirationTimeMicros)
                 .setCompletion((o, e) -> {
                     if (e != null) {
-                        sendSelfFailurePatch(task, e.getMessage());
+                        if (failureHandler != null) {
+                            failureHandler.accept(e);
+                        }
                         return;
                     }
 
-                    NodeSelectorService.SelectOwnerResponse rsp = o.getBody(
-                            NodeSelectorService.SelectOwnerResponse.class);
+                    SelectOwnerResponse rsp = o.getBody(SelectOwnerResponse.class);
 
                     if (!rsp.isLocalHostOwner) {
-                        logWarning("Current node %s is no longer owner for the factory %s. Cancelling synchronization",
-                                this.getHost().getId(), task.factorySelfLink);
-
-                        sendSelfCancellationPatch(task, "Local node is no longer owner for this factory.");
+                        if (nonOwnerHandler != null) {
+                            nonOwnerHandler.accept(rsp);
+                        }
                         return;
                     }
 
-                    // Recursively call handleSynchronizeStage
-                    // without owner verification.
-                    handleSynchronizeStage(task, false);
+                    if (ownerHandler != null) {
+                        ownerHandler.accept(rsp);
+                    }
                 });
 
         getHost().selectOwner(task.nodeSelectorLink, task.factorySelfLink, selectOp);
-        return true;
     }
 
     private void synchronizeService(State task, String link, Operation.CompletionHandler c) {
@@ -824,9 +942,51 @@ public class SynchronizationTaskService
                         return;
                     }
 
+                    if (this.parent != null && this.parent.hasChildOption(ServiceOption.PERSISTENCE)
+                            && this.isCheckpointEnabled) {
+                        Consumer<SelectOwnerResponse> ownerHandler = (selectOwnerResponse) -> {
+                            if (selectOwnerResponse.availableNodeCount > 1) {
+                                createCheckpointsAndReschedule(task);
+                            } else {
+                                sendSelfFinishedPatch(task);
+                            }
+                        };
+                        Consumer<SelectOwnerResponse> nonOwnerHandler = (selectOwnerResponse) -> {
+                            sendSelfFinishedPatch(task);
+                        };
+                        Consumer<Throwable> failureHandler = (ex) -> {
+                            sendSelfFinishedPatch(task);
+                        };
+                        verifySynchronizationOwnership(task, ownerHandler, nonOwnerHandler, failureHandler);
+                        return;
+                    }
+
                     sendSelfFinishedPatch(task);
                 });
         sendRequest(getNodeSelectorStateOp);
+    }
+
+    private void createCheckpointsAndReschedule(State task) {
+        CheckpointService.CheckpointState s = new CheckpointService.CheckpointState();
+        s.timestamp = task.startTimeMicros;
+        s.factoryLink = this.parent.getSelfLink();
+
+        logInfo("Creating checkpoints for factory %s with timestamp: %d", s.factoryLink, s.timestamp);
+        Operation post = Operation.createPost(UriUtils.buildUri(this.getHost(), CheckpointService.FACTORY_LINK))
+                .setBody(s)
+                .setReferer(this.getUri())
+                .setCompletion((op, ex) -> {
+                    getHost().scheduleCore(() -> {
+                        SynchronizationTaskService.State scheduleTask = this.parent.createSynchronizationTaskState(
+                                task.membershipUpdateTimeMicros);
+                        Operation.createPost(this, ServiceUriPaths.SYNCHRONIZATION_TASKS)
+                                .setBody(scheduleTask)
+                                .sendWith(this);
+                    }, this.schedulePeriodSeconds, TimeUnit.SECONDS);
+
+                    sendSelfFinishedPatch(task);
+                });
+        this.getHost().broadcastRequest(this.parent.getPeerNodeSelectorPath(), CheckpointService.FACTORY_LINK, false, post);
     }
 
     private void setFactoryAvailability(
@@ -854,6 +1014,10 @@ public class SynchronizationTaskService
                     }
                 });
         sendRequest(put);
+    }
+
+    public void setParentService(FactoryService factoryService) {
+        this.parent = factoryService;
     }
 
     @Override